chat_backend/requirements.txt
dnspython==2.4.2
einops==0.8.0
eventlet==0.33.3
Flask==2.2.3
Flask-SocketIO==5.3.3
gunicorn==21.2.0
nltk==3.8.1
numpy==1.26.2
openai==1.56.0
pandas==2.1.3
psycopg2-binary==2.9.9
pydantic==2.5.2
pymssql==2.3.0
pyodbc==4.0.34
python-dotenv==1.0.0
python-socketio==5.10.0
sentence-transformers==2.7.0
simple-websocket==1.0.0
snowflake-connector-python==3.6.0
sqlparse==0.4.4
torch==2.1.2
tqdm==4.66.1
Werkzeug==2.3.7

# engine
SQLAlchemy==2.0.23
databricks-sql-connector==3.0.1
ibm_db==3.2.3
